test(anyharness): split terminal command parser tests
